:recycle: refactor: reactoring code
