+++ kok-api/src/main/java/com/kok/kokapi/room/adapter/in/dto/request/CreateRoomRequest.java
@@ -1,4 +1,4 @@
-package com.kok.kokapi.room.adapter.in.dto;
+package com.kok.kokapi.room.adapter.in.dto.request;
 
 
 import jakarta.validation.constraints.Min;
