+++ kok-api/src/main/java/com/kok/kokapi/room/adapter/in/web/RoomController.java
@@ -1,6 +1,8 @@
 package com.kok.kokapi.room.adapter.in.web;
 
-import com.kok.kokapi.room.adapter.in.dto.CreateRoomRequest;
+import com.kok.kokapi.common.response.ApiResponseDto;
+import com.kok.kokapi.room.adapter.in.dto.request.CreateRoomRequest;
+import com.kok.kokapi.room.adapter.in.dto.response.RoomResponse;
 import com.kok.kokcore.domain.Room;
 import com.kok.kokcore.usecase.CreateRoomUseCase;
 import jakarta.validation.Valid;
@@ -19,13 +21,15 @@ public class RoomController {
     private final CreateRoomUseCase createRoomUseCase;
 
     @PostMapping("/create")
-    public ResponseEntity<Room> createRoom(@Valid @RequestBody CreateRoomRequest request) {
+    public ResponseEntity<ApiResponseDto<RoomResponse>> createRoom(@Valid @RequestBody CreateRoomRequest request) {
         Room room = createRoomUseCase.createRoom(
                 request.getRoomName(),
                 request.getCapacity(),
                 request.getHostProfile(),
                 request.getPassword()
         );
-        return ResponseEntity.ok(room);
+
+        var response = RoomResponse.from(room);
+        return ResponseEntity.ok(ApiResponseDto.success(response));
     }
 }
