+++ kok-api/src/main/java/com/kok/kokapi/room/application/service/RoomService.java
@@ -1,6 +1,6 @@
 package com.kok.kokapi.room.application.service;
 
-import com.kok.kokcore.application.port.out.RoomRepository;
+import com.kok.kokcore.application.port.out.SaveRoomPort;
 import com.kok.kokcore.domain.Room;
 import com.kok.kokcore.usecase.CreateRoomUseCase;
 import lombok.RequiredArgsConstructor;
@@ -10,11 +10,11 @@ import org.springframework.stereotype.Service;
 @RequiredArgsConstructor
 public class RoomService implements CreateRoomUseCase {
 
-    private final RoomRepository roomRepository;
+    private final SaveRoomPort saveRoomPort;
 
     @Override
     public Room createRoom(String roomName, int capacity, String hostProfile, String password) {
         Room room = Room.create(roomName, capacity, hostProfile, password);
-        return roomRepository.save(room);
+        return saveRoomPort.save(room);
     }
 }
